Add bytecode writing; Change floats from `long double` -> `double` to fix alignment issues when writing - long double is apparently padded to 16 bytes.

+++ include/AST/Stmts.hpp
@@ -107,7 +107,7 @@ public:
 
     inline String getMoveLexDataStr() { return val->getMoveDataStr(); }
     inline StringRef getLexDataStr() const { return val->getDataStr(); }
-    inline long double getLexDataFlt() const { return val->getDataFlt(); }
+    inline double getLexDataFlt() const { return val->getDataFlt(); }
     inline int64_t getLexDataInt() const { return val->getDataInt(); }
 };
